Break out microphone worker into its own file
